First version of the logger library, with unit tests for core classes
Tests are not yet complete but cover main functionalities of the
core classes of the library.

The JSON layout needs to be updated and the library thoroughly tested
in real environments.

To build the library pending a proper build mechanism:
npm install
node node_modules/requirejs/bin/r.js -o name=woodman-browser
  baseUrl=lib/ out=foo.js preserveLicenseComments=false