Life cycle (start/stop) added to logger context and appenders
Initialization is a two-step process:
1. the configuration is applied, all appenders are created
2. the appenders are started in parallel.

The second step may be asynchronous as appenders may need to
perform I/O operations to set things up. The start/stop functions
call the callback function they receive as parameter once done.

From an external point of view, the Woodman's library should be
started with a call to "load" as in:

  Woodman.load(config, function (err) {
    // App code goes here
  });