Additivity: parent trace level was checked. It should not.
Additivity rules say that the append function of a logger is to
call that of the logger's parent when the logger is additive, up
until a non additive logger is found or up until the process reaches
the root logger.

The trace level of the parent does not matter in these rules. The
code was not calling the appenders of the parent logger if its trace
level was not the right one for the log event. This update fixes the
problem. This mechanism allows one to define a root logger that logs
"error" message except for a module "foo" for which we want the root
logger to log "info" messages.