Added distribution of Woodman that runs in browsers and node.js
The dist/woodman.js contains all known appenders and layouts. It
runs in Web browsers as well as in node.js environments, using an
AMD loader or not.

Examples that use this distribution:
 examples/browser/standalone.html
 examples/browser/amd.html
 node examples/node.js/standalone.js
 node examples/node.js/amd.js

NB: there is no reliable way to determine whether some code is run
in a Web browser, server-side or somewhere else, using AMD or not,
because all global variables may be re-defined in JavaScript. The
code handles most common cases.