SocketAppender defined. JSONLayout fixed. Version 0.1.2
The SocketAppender uses socket.io to send logs in real time to some
server endpoint. The server needs to be up and running for the
appender to work well (otherwise "start" may never return for the
being, to be improved later on).

An example socket server is provided in examples/socket.io/server.js