Precompiler: improved "comment" option to comment out lines properly
The "comment" precompilation option tells the precompiler to comment
out the lines instead of removing them (when possible, note some
updates cannot easily be commented because they are made within an
expression statement). The precompiler used to wrap the code in
"/* ... */" but that breaks when the code already contains the "*/"
string.

The update uses "//" at the beginning of all lines that need to be
commented.