Moved "devDependencies" to "dependencies" to use precompiler from app
When Woodman is included in another app with an npm install, the
"devDependencies" are not installed. That's good and logic. Problem
arises when the app actually wants to use the precompiler that comes
with Woodman to build the code: the dev dependencies are nowhere to
be found in that case.

Moving the dev dependencies to regular dependencies solves the
problem. It does come with the additional cost that Woodman now
seems to depend on many different libraries whereas these libraries
are only really necessary to remove the references to Woodman during
compilation. Splitting the library and the precompiler into two
packages would perhaps be a better solution in the future.