Precompiler: fixed exception on indirect calls and "load" replacement
The code that detected whether the code contained an indirect call
such as "require('woodman').getLogger()" was not checking the
"woodman" name and thus was throwing the exception in other cases
that had nothing to do with Woodman.

Similarly, the code that replaced code such as "woodman.load()" was
using a regular expression to detect the end of the first parameter
(the config object). That's intrinsically wrong since there is no
proper way to detect the end of the first parameter with a regular
expression. Now that the AST tree is fully extended with helpers such
as the "source" method, it's easy to extract the second parameter of
the load call and wrap it into "(param)()" to execute the function
directly.