Precompiler: now using "require" in deps to replace "woodman"
To remove the dependencies to Woodman in "define" calls, the
precompiler replaced them with an empty string, hoping that require.js
would simply load an empty module in such cases. That does not
work as expected. The "empty:" value that require.js proposes in its
doc would be great but only seems to work in the "paths" property of
require.js config.

The best would be to remove the dependency altogether, but that's not
so easy to do from an AST tree perspective.

All in all, the only module that always exists in an AMD environment
is "require", so that's what the precompiler uses now.

For instance:
 define(['woodman'], function (woodman) {});
becomes:
 define(['require'], function (woodman) {});