Switched to require.js "shortcut" (a.k.a "sugar") notation
Modules used to be defined as "regular" require.js modules, with
dependencies listed upfront, as in:

 define(['blah', 'foo'], function (Blah, Foo) {});

They are now defined using require.js "shortcut" notation:

 define(function (require) {
   var Blah = require('blah');
   var Foo = require('foo');
 });

... which has the merit to be much more readable and to ease detection
of unused dependencies and possible mismatches.

Internally, require.js uses Function.prototype.toString() to parse
the code beforehand and extract dependencies, in order to be able
to load these dependencies asynchronously. This should work in all
sensible environment (PS3 and old Opera mobile versions excepted
according to require.js). Woodman is likely to be use as a compiled
library in any case, and require.js optimizer takes care of converting
that notation into the appropriate one when it optimizes things away.