Log event filtering added. Filters not yet initialized. Not tested
Filters allow log events to be evaluated to determine whether they
should be published. Filters are typically useful to implement more
complex filtering rules than a filter on the logger's name and/or
on the event level. A typical example would be the use of a regular
expression to filter out messages of a certain kind.

Filters may apply context-wide, within loggers or within appenders.
For more info, see:
http://logging.apache.org/log4j/2.x/manual/filters.html

Woodman does not support appender reference filters for the time
being.

Filters still need to be created by LoggerContext during
initialization.

Code is most likely not operational in any case as it has not been
tested yet.