Filters created during initialization. Ready for testing
Filters referenced in the configuration are created when Woodman is
initialized. The configuration may reference context-wide filters,
logger filters or appender filters. Each configuration may either be
an array or an object indexed by the filter's type.

Example of a context-wide filter using an array:
  {
    "filters": [
      {
        "type": "RegexFilter",
        "regex": "Hello"
      }
    ],
    "appenders": [...],
    "loggers": [...]
  }

Same example with an object:
  {
    "filters": {
      "RegexFilter": {
        "regex": "Hello"
      }
    },
    "appenders": [...],
    "loggers": [...]
  }

Code is ready for testing but has not been tested yet.