FileAppender implemented.
Note that the need to "require" the File System module is a bit of a
pain because the main distribution of Woodman uses Almond which
overrides the "require" variable if it exists and does not support
loading Node.js modules.

Current implementation uses a global "nodefs" variable but that is not
entirely satisfactory. See comments and TODO in FileAppender for
some possible solution.