Fixed Node.js builds. FileAppender uses "rootRequire" to load "fs"
The FileAppender, which may only run in a Node.js runtime environment
due to its dependency on "fs" now attempts to load the File System
module using "rootRequire" if that variable is set or "require" if
not. The "rootRequire" is intended to point to Node.js's own "require"
method even when Woodman is built with Almond (Almond defines its own
version of "require" that does not support Node.js modules).

Note that that code hides the fact that it may issue a call to
"require('fs')" so that Require.js optimizer misses it (it would not
know how to include the File System module in a build).

Fixed the different distributions accordingly. The distributions
intended for use in a Node.js environment depended on "amdefine"
but that is not needed and kind of negates the claim that the
standalone distribution of Woodman is standalone.

New woodman-node-amd distribution created that does what its name
suggests (Node.js distribution that exposes Woodman as an AMD module).