Add a new RollingFileAppender. This appender is like a FileAppender, but allows to specify either a size limit or a max age limit.
The limits are defined in the configuration, under 'triggeringPolicy'. If
'size' is specified, it is the maximum number of bytes that a log file can
grow up to. If 'time' is specified, it is, in seconds, the maximum time we
use a log file before a new one is created.