Documentation: introduction updated to look less scary
While it remains a good idea to mention the different features that
Woodman exposes, it seemed even better to start by mentioning that
using Woodman is really just as simple as using the "console" object.
The complex part in Woodman is the configuration mechanism, but users
do not necessarily need to understand that mechanism to start using
Woodman.