Added Woodman traces to precompiler library to trace execution
The traces are not enabled by default. They should be useful to
track down the steps that take a lot of time and might be worth
improving.

Not surprisingly, the operation that seems to consume the most amount
of time is parsing the JavaScript code into an AST tree (with the
"range" flag set to keep track of the source code).