Cleaned precompiler cli code and options. New --verbose option.
The precompiler command-line interface now displays a friendly
usage message when called with "--help", "-h" or with invalid
parameters and options.

New "--verbose" option added that outputs Woodman traces to the
console to follow the execution of the precompiler and track
performance issues.

Note the new cli is not backward-compatible with the previous one
as the "--keep" option was named "-level" previously. Upgrade should
be pretty straightforward though.