Layout and Message updated for "appendStrings: false" option (#5)
The goal is for the console appender to send objects to the console
that are not serialized as strings when so requested with an
"appendStrings: false" option.

Internally, the appender will call "Layout.toMessageBits" to retrieve
bits of messages to send to the console. The layout delegates to
"Message.getFormattedParams" to create the appropriate bits.

JSONLayout and PatternLayout now need to implement "toMessageBits"
and "ConsoleAppender" still needs to call the "Layout.toMessageBits"
function.