Config shortcut "console" now preserves objects in Web browsers (#5)
Loading Woodman with the "console" configuration shortcut in a Web
browser now behaves as the usual console for objects, meaning it
logs them as objects that can be expanded with the console.

Note objects are not preserved when Woodman is run in a node.js
environment. Also note that the code is a bit dumb for the time being
in that it assumes that not running in a node.js environment means
running in a Web browser whereas there are other JS runtimes around.