Log event handling logic split between Logger and LoggerContext (#3)
The logger now creates the LogEvent instance and sends it to the
underlying LoggerContext for processing. In turn, the LoggerContext
will typically just give it back to the Logger, provided that the
context-wide filter does not reject the event and provided that
Woodman is indeed started.

That's a first step towards being able to cache events in memory while
Woodman is being started.