Log events kept in memory while Woodman starts (#3)
The logger context now keeps log events it receives before the context
is up and running in memory. It processes the events as soon as the
context is operational.

To avoid out of memory issues, the context only keeps the last 1000
most recent log events in memory. It logs a warning when messages had
to be discarded.

The maximum number of log events kept in memory may be adjusted
through the "maxPendingEvents" configuration setting. Beware though:
by definition, this setting can only be taken into account when
Woodman is initialized!