Tests: switched to RequireJS shortcut notation. BaseURL defined.
The shortcut notation is now used consistently across Woodman.

Setting RequireJS "baseURL" in the test runner is actually a good
idea because RequireJS does not know how to normalize paths that
go above the folder that contains the test runner by default and,
while it correctly resolves modules in the end, it fills the modules
array with names such as:
 spec/../../lib/./loglevel
 spec/../../lib/././loglevel

... which has the annoying drawback that module variables that should
typically be defined once and for all are actually defined once per
entry in the modules array.

Setting the baseUrl to the root folder solves the problem.