Precompiler crashed in optimized script with sugar syntax (#25)
The precompiler took for granted that there was always a match
between the number of dependencies listed in the call to "define"
and the number of parameters in the following function.

That is typically NOT the case when the sugar syntax is used and
when the code is optimized with RequireJS optimizer as the code
produced looks like:

define('name', ['require', 'woodman'], function (require) {
  var woodman = require('woodman');
});

Regression tests defined.