Fixed distributions to expose AMD modules that r.js can pick up (#20)
Full Woodman distributions worked correctly but crashed if they were
included in code built with RequireJS r.js optimizer.

This happened because RequireJS r.js optimized failed to detect the
call to "define" that exposed the distribution as an AMD module
because it was hidden too deep in the code.

The call to "define" now appears at the root of the distribution
(provided that "define" is set, that is).