Added "%domain" pattern to PatternLayout (Node.js env only) (#28)
The new "%domain" pattern tells Woodman to check the active Node.js
domain and to output the ID of that domain if it is defined. Domains
are typically useful to track the execution of an HTTP request in an
HTTP server throughout the code.

The ID of the domain is taken to be defined in the "id" property of
the active domain. One may specify another possibly nested property
to use in the pattern parameter as in "%domain{nested.name.id}",
which would be able to extract the ID from a domain structure such as:
 {
   nested: {
     name: {
       id: 'my domain'
     }
   }
 }